refactor: 온보딩 관련 스캐폴드 패키지 구조 변경 및 UI 스타일 수정
- `OnboardingScaffold` 및 `ProgressBarScaffold` 위치를 `signup.scaffold` 패키지로 이동
- `TermsDetailScreen`, `OnboardingTermsScreen` 등 온보딩 화면 내 텍스트 컬러를 `gray9` 또는 `black`으로 조정하여 가독성 개선
- `LoginScreen` 구글 로그인 텍스트 및 `AfternotePopupCardLayout` 본문 텍스트에 누락된 컬러 명시
- `OnboardingProfileScreen`에서 `typography.h1`에 중복 적용된 `FontWeight.Bold` 제거 및 미리보기 설정 변경
- 패키지 이동에 따른 프로젝트 내 임포트 경로 업데이트
