+++ feature/onboarding/presentation/src/main/java/com/afternote/feature/onboarding/presentation/login/LoginScreen.kt
@@ -253,6 +253,7 @@ private fun SocialLoginGroup(
             Text(
                 text = stringResource(R.string.login_google),
                 style = AfternoteDesign.typography.captionLargeB,
+                color = AfternoteDesign.colors.gray8,
             )
             Spacer(Modifier.weight(1f))
         }
