+++ feature/onboarding/presentation/src/main/java/com/afternote/feature/onboarding/presentation/signup/SignUpPasswordScreen.kt
@@ -26,7 +26,7 @@ import com.afternote.core.ui.Label
 import com.afternote.core.ui.theme.AfternoteDesign
 import com.afternote.core.ui.theme.AfternoteTheme
 import com.afternote.feature.onboarding.presentation.R
-import com.afternote.feature.onboarding.presentation.scaffold.ProgressBarScaffold
+import com.afternote.feature.onboarding.presentation.signup.scaffold.ProgressBarScaffold
 
 @Composable
 fun SignUpPasswordScreen(
